+++ src/main/java/domain/type/Color.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.type;
 
 public enum Color {
     WHITE, BLACK
